requirements.txt
instagrapi
Pillow
Flask
Werkzeug
cryptography